atomisp2: css2400 for irci_master_20150303_1500
Merge CSS2.1 for css2400  for
irci_master_20150303_1500 firmware

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-24134
Reviewed-on: https://icggerrit.ir.intel.com/44559
Build: viefandr
Integrator: Wei Tao <wei.tao@intel.com>
Approver: Wei Tao <wei.tao@intel.com>
Build: Wei Tao <wei.tao@intel.com>
Maintainer: Wei Tao <wei.tao@intel.com>

+++ drivers/external_drivers/camera/drivers/media/pci/atomisp2/css2400b0_v21/isp/kernels/raw/raw_1.0/ia_css_raw_param.h
@@ -30,8 +30,8 @@ struct sh_css_isp_raw_isp_config {
 	uint32_t two_ppc;
 	uint32_t stream_format; /* enum sh_stream_format */
 	uint32_t deinterleaved;
-	uint8_t start_column; /*left crop offset*/
-	uint8_t start_line; /*top crop offset*/
+	uint32_t start_column; /*left crop offset*/
+	uint32_t start_line; /*top crop offset*/
 	uint8_t enable_left_padding; /*need this for multiple binary case*/
 };
